requirements.txt
pytest
pytest-html
pytest-rerunfailures
playwright
python-dotenv